% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_px_file.R
\name{download_px_cube}
\alias{download_px_cube}
\title{Download px cube file for px id into a local directory}
\usage{
download_px_cube(px_id, local_dir = tempdir())
}
\arguments{
\item{px_id}{px identification number on the fso website}

\item{local_dir}{optional defaults to tempdir()}
}
\value{
download_path
}
\description{
Download px cube file for px id into a local directory
}
\examples{
download_px_cube('px-x-0102020207_102', '/tmp')
          download_px_cube('px-x-0102020207_102')
}
