% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattab-overview.R
\name{enrich_fso_px_cube_list}
\alias{enrich_fso_px_cube_list}
\title{enrich px data cubes}
\usage{
enrich_fso_px_cube_list(file_path, lang = "de", start = 1, end = 1)
}
\arguments{
\item{file_path}{csv file path for input and output}

\item{lang}{language fot the list}

\item{start}{index to start processing}

\item{end}{index to end processing}
}
\value{
csv file path of output file
}
\description{
enrich px data cubes
}
\examples{
enrich_fso_px_cube_list("output.csv", lang = "en", start=1, end=1)
}
