% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattab-overview.R
\name{get_fso_px_cube_list}
\alias{get_fso_px_cube_list}
\title{make fso overview list on px data cubes}
\usage{
get_fso_px_cube_list(lang = "de", output_dir = getwd(), force = FALSE)
}
\arguments{
\item{lang}{language fot the list}

\item{output_dir}{where to write the csv output file}

\item{force}{create file even if it already exists}
}
\value{
csv file path of output file
}
\description{
make fso overview list on px data cubes
}
\examples{
get_fso_px_cube_list(lang = "en", force = TRUE)
}
