% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_px_url.R
\name{get_fso_px_download_url}
\alias{get_fso_px_download_url}
\title{Get pxt download url for px id}
\usage{
get_fso_px_download_url(px_id)
}
\arguments{
\item{px_id}{identifier for px cube}
}
\value{
download_url
}
\description{
Get pxt download url for px id
}
\examples{
get_fso_px_download_url('px-x-0102020207_102')
}
