% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsoPxLoc.R
\name{get_px_download_url}
\alias{get_px_download_url}
\title{Get Download Url for px id}
\usage{
get_px_download_url(px_id)
}
\arguments{
\item{px_id}{}
}
\value{
download_url
}
\description{
Get Download Url for px id
}
\examples{
get_asset_id('px-x-0102020204_102')
}
